requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
urllib3==2.1.0
